#8, Refactor package.
1. move class `JetPackFunction` and `User` to package `bean`;
2. move class `SharedPreferencesUtil` to package `shared_preference`;
3. delete empty package `util`;
