+++ app/src/main/java/com/imeepwni/jetpack/main/FunctionAdapter.kt
@@ -2,7 +2,7 @@ package com.imeepwni.jetpack.main
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.imeepwni.jetpack.data.JetPackFunction
+import com.imeepwni.jetpack.data.bean.JetPackFunction
 
 /**
  * 作者：Created by guofeng on 2019/10/10
